Move result to separate file
Fix result
